refactor(plugin): Merge MiAZAIChat into MiAZAIAssistant
