--- data/resources/plugins/MiAZAIChat/miazaic/prompt.py
@@ -1,12 +1,0 @@
-#!/usr/bin/python3
-
-
-def system_prompt(document_name: str) -> str:
-    return (
-        "You are a helpful assistant answering questions about a single "
-        "document.\n"
-        f"The document is named '{document_name}'.\n"
-        "Answer using only the content of that document. If the answer is not "
-        "in the document, say clearly that you cannot find it. Be concise and "
-        "quote the relevant part of the document when it helps."
-    )
